pushing

+++ srcs/fdf.h
@@ -6,7 +6,7 @@
 /*   By: wfung <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/04/26 19:06:47 by wfung             #+#    #+#             */
-/*   Updated: 2017/05/18 10:09:03 by wfung            ###   ########.fr       */
+/*   Updated: 2017/05/18 17:15:20 by wfung            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,12 +40,12 @@ typedef struct				s_fdfstore
 	int						z;		//find out how to use
 	int						col;
 	int						row;
-	int						win_x;		//save max x axis of window
-	int						win_y;		//save max y axis of window
-	int						center_x;
-	int						center_y;
-	int						start_x;
-	int						start_y;
+	float					win_x;		//save max x axis of window
+	float					win_y;		//save max y axis of window
+	float					center_x;
+	float					center_y;
+	float					start_x;
+	float					start_y;
 }							t_fdfstore;
 
 void	ft_make_intarray(t_fdfstore *grid);
